requirements.txt
python-telegram-bot
nest_asyncio
licensing
paramiko
asyncssh